requirements.txt
aws-cdk-lib==2.73.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.73.0a0
